requirements.txt
aenum==3.1.15
bleak==0.20.2
dnspython==2.3.0
eventlet==0.36.1
Flask_Cors==4.0.0
Flask_SocketIO==5.3.6
Flask==2.2.5
MarkupSafe==2.1.5
mock==5.1.0
packaging==21.3
python-dateutil==2.9.0.post0
requests==2.31.0
ruamel.yaml==0.18.6
webargs==8.3.0
Werkzeug==2.2.3
paho-mqtt==1.6.1